Move globals components to app
