+++ app/_components/globals/Hero.tsx
