+++ app/products/[slug]/_components/product-detail/ProductImageGallery.tsx
