+++ app/products/[slug]/_components/product-detail/ProductInfo/ProductCategories.tsx
