+++ app/products/[slug]/_components/product-detail/ProductInfo/ProductSocialMedia.tsx
